Release Wigii 4.610 G176 Author: lionel.weber Date: 24.04.2018 17:07
Message:
Added a new parameter for listView field: displayHTMLinLabel 0/1 allow to display some html code in a field list label if needed.
Added a new parameter defaultEmailingFrom. If defined then the Emailing form will by default use this one for the from.
Do not display the max number of subscription when the max is reached on a public form. Only keep a generic maximum is reached message.
Correct the background-color for public form to be applied to body and not only to html (which made it almost never visible).
Add new icon with a file and a green check mark on it.