Release Wigii 4.614 G189 Author: lionel.weber Date: 18.05.2018 14:00
Message:
Add a new parameter: allowAudioRecording on field with datatype Files.
If allowAudioRecording then an audio recording button is available above the file field. When pressing on it the user can record from his mic. The audio file is then transfered as a file to the normal WigiiFormExecutor flow.
Enhanced the audio and video standard player by maximizing the usage of html5 technologies.