Release 4.714 G397, graphical, functional and technical release
Enhance UI by adding module icons
Correct bootstrap css preventing checkbox or radiobutton labels to be bold.

Add the recurring activity. This activity can be activated anywhere using recurringJsCodeInElementDetail in a jsCodeInDetailExp of an item having a timeRange or date field.

Updated the ElementService with the new method insertElementCopy allowing to insert new element with data of an existing element.
Added in the RecordEvaluator ability to test if RootPrincipal is available using getRootPrincipalIfAvailable or assertRootPrincipalIsAvailable.
Mapped the cs_g funcExp to php function.
Corrected FormExecutor to not execute InDetailExp if not in DetailElementFormExecutor or InFormExp if not in EditElementFormExecutor

Enhanced the WigiiApi.js with additionnal methods:
elementIsEditable, Returns if current element has the modify button active
currentGroupId, Returns the ID of the current group
writableCurrentGroupId, Returns the ID of the current group if group can be written